# fix performance issue in torch.sparse.mm reduce mode (#94969)